+++ tests/client/test_get_press_summaries_for_document_uri.py
@@ -19,25 +19,23 @@ class TestGetPressSummariesForDocumentUri(TestCase):
         mock_press_summary,
     ):
         mock_eval.return_value = "EVAL"
-        mock_get_marklogic_response.return_value = ["/foo/bar/baz/1", "/foo/bar/baz/2"]
+        mock_get_marklogic_response.return_value = ["foo/bar/baz/1", "foo/bar/baz/2"]
 
-        for uri in ["foo/bar", "/foo/bar"]:
-            with self.subTest(uri=uri):
-                self.client.get_press_summaries_for_document_uri(DocumentURIString(uri))
+        self.client.get_press_summaries_for_document_uri(DocumentURIString("foo/bar"))
 
-                mock_get_marklogic_response.assert_called_with("EVAL")
-                mock_eval.assert_called_with(
-                    {
-                        "parent_uri": "/foo/bar",
-                        "component": "pressSummary",
-                    },
-                    "get_components_for_document.xqy",
-                )
+        mock_get_marklogic_response.assert_called_with("EVAL")
+        mock_eval.assert_called_with(
+            {
+                "parent_uri": "foo/bar",
+                "component": "pressSummary",
+            },
+            "get_components_for_document.xqy",
+        )
 
-                mock_press_summary.assert_has_calls(
-                    [
-                        call("/foo/bar/baz/1", self.client),
-                        call("/foo/bar/baz/2", self.client),
-                    ],
-                    any_order=True,
-                )
+        mock_press_summary.assert_has_calls(
+            [
+                call("foo/bar/baz/1", self.client),
+                call("foo/bar/baz/2", self.client),
+            ],
+            any_order=True,
+        )